finished updating the org-management backend
I completed the add-db, update-db and remove-db endpoints and added 2 more for get-members, and get-dbs.

I didn't like the idea of passing the db_info (which is JSON) as plain text so the endpoints that need the db_info require you to send the db_info as a JWT string which is signed with the supabase jwt secret.